+++ lib/src/formWidget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:simple_form_builder/global/constant.dart';
 import 'package:simple_form_builder/src/widgets/customDropdownWidget.dart';
 
-import 'models/checklistModel.dart';
+import '../global/checklistModel.dart';
 
 class FormBuilder extends StatefulWidget {
   final Map<String, dynamic> initialData;
